+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## Changelog
 
-### 12.46.1
+### 12.47.0
 
 * `[fsutil]` Added bufio writer for copying files
+* `[knf]` Added method `Is` for checking property value
 * `[options]` Added method `Is` for checking option value
 * `[fsutil]` Fixed bug with changing file mode for copied file if custom umask is used
 * `[options]` Code refactoring